Back to uint16

+++ Source/Voxel/Voxel_Chunk.h
@@ -46,7 +46,7 @@ public:
 
 	//The version of the chunk replicated to clients
 	UPROPERTY(ReplicatedUsing = OnRep_NetworkData)
-		uint8 NetworkData[999];
+		uint16 NetworkData[999];
 		//TArray<uint16> NetworkData;
 
 	//Blocks stored in an easy to use 3d array
